implemented data source methods for tableView.

+++ Documents/Documents/NSString+WordCount.h
@@ -15,7 +15,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface NSString (WordCount)
 
-- (NSInteger)checkWordCount:(NSString *)string;
+- (int)checkWordCount:(NSString *)string;
 
 @end
